@@ -1,79 +1,106 @@
 require 'active_support/inflector'
 
 module Api
-
   module Generator
     def self.generator_class(edition)
       "Api::Generator::#{edition.container.class.to_s}".constantize
     end
 
     def self.edition_to_hash(edition, *args)
-      generator_class(edition).edition_to_hash(edition, *args)
+      generator = generator_class(edition)
+      publication_fields =  [:slug, :updated_at]
+      edition_fields     =  [:title, :alternative_title, :overview] + generator.extra_fields
+
+      attrs = edition.container.as_json(:only => publication_fields)
+      attrs.merge!(edition.as_json(:only => edition_fields))
+
+      if edition.respond_to?(:parts)
+        non_blank_parts = []
+        edition.parts.each do |part|
+          non_blank_parts << part unless part.body.blank?
+        end
+        attrs['parts'] = non_blank_parts.sort_by(&:order).collect { |p| p.as_json(:only => [:slug, :title, :body]) }
+      end
+
+      if edition.respond_to?(:expectations)
+        attrs['expectations'] = edition.expectations.map {|e| e.as_json(:only => [:css_class,:text]) }
+      end
+
+      attrs['type'] = edition.container.class.to_s.underscore
+      generator.edition_to_hash(attrs, edition,*args)
     end
 
-     module Guide
-      def self.edition_to_hash(edition,options={})
-        attrs = edition.guide.as_json(:only => [:audiences, :slug, :tags, :updated_at, :section, :related_items])
-        attrs.merge!(edition.as_json(:only => [:title, :alternative_title, :overview]))
-        attrs['parts'] = edition.parts.sort_by(&:order).collect { |p| p.as_json(:only => [:slug, :title, :body]).merge('number' => p.order) }
-        attrs['type'] = 'guide'
+    class Base
+      def self.extra_fields
+        []
+      end
+
+      def self.edition_to_hash(attrs, guide, options={})
         attrs
       end
     end
 
-    module Answer
-      def self.edition_to_hash(edition,options={})
-        attrs = edition.answer.as_json(:only => [:audiences, :slug, :tags, :updated_at, :section, :related_items])
-        attrs['type'] = 'answer'
-        attrs.merge!(edition.as_json(:only => [:title,:body, :alternative_title, :overview]))
-      end
-    end
-    
-    module Transaction
-      def self.edition_to_hash(edition,options={})
-        attrs = edition.transaction.as_json(:only => [:audiences, :slug, :tags, :updated_at, :section, :related_items])
-        attrs['type'] = 'transaction'
-        attrs['expectations'] = edition.expectations.map {|e| e.as_json(:only => [:css_class,:text]) }
-        attrs.merge!(edition.as_json(:only => [:title, :introduction, :more_information, :will_continue_on,:link, :alternative_title, :overview, :minutes_to_complete, :uses_government_gateway]))
+    class Guide < Base
+      def self.extra_fields
+        [ :video_url,
+          :video_summary]
       end
     end
 
-    module LocalTransaction
-      def self.edition_to_hash(edition, options = {})
+    class Programme < Base
+    end
+
+    class Answer < Base
+      def self.extra_fields
+        [:body]
+      end
+    end
+
+    class Transaction < Base
+      def self.extra_fields
+        [ :introduction,
+          :more_information,
+          :will_continue_on,
+          :link,
+          :minutes_to_complete,
+          :uses_government_gateway,
+          :alternate_methods]
+      end
+    end
+
+    class LocalTransaction < Base
+      def self.extra_fields
+        [ :introduction,
+          :more_information,
+          :minutes_to_complete]
+      end
+
+      def self.authority_to_json(la)
+        la.as_json(:only => [:snac, :name], :include => {:lgils => {:only => [:url, :code]}})
+      end
+
+      def self.edition_to_hash(attrs, edition, options = {})
         snac = options[:snac]
         all  = options[:all]
-        attrs = edition.local_transaction.as_json(:only => [:audiences, :slug, :tags, :updated_at, :section, :related_items])
-        attrs.merge!(edition.as_json(:only => [:title, :introduction, :more_information, :alternative_title, :overview, :minutes_to_complete]))
-        attrs['type'] = 'local_transaction'
-        attrs['expectations'] = edition.expectations.map { |e| e.as_json(:only => [:css_class, :text]) }
-        if snac
-          attrs['authority'] = edition.local_transaction.lgsl.authorities.where(snac: snac).first.as_json(:only => [:snac, :name], :include => {:lgils => {:only => [:url, :code]}})
-        elsif all
-          attrs['authorities'] = edition.local_transaction.lgsl.authorities.all.as_json(:only => [:snac, :name], :include => {:lgils => {:only => [:url, :code]}})
+        if snac || all
+          las = edition.local_transaction.lgsl.authorities
+          if snac
+            attrs['authority'] = authority_to_json(las.where(snac: snac).first)
+          elsif all
+            attrs['authorities'] = authority_to_json(las.all)
+          end
         end
         attrs
       end
     end
 
-    module Place
-      def self.edition_to_hash(edition, options={})
-        attrs = edition.place.as_json(:only => [:audiences, :slug, :tags, :updated_at, :section, :related_items])
-        attrs.merge!(edition.as_json(:only => [:title, :introduction, :more_information, :place_type, :alternative_title, :overview]))
-        attrs['expectations'] = edition.expectations.map {|e| e.as_json(:only => [:css_class,:text]) }
-        attrs['type'] = 'place'
-        attrs
+    class Place < Base
+      def self.extra_fields
+        [ :introduction,
+          :more_information,
+          :place_type]
       end
-    end
-    
-    module Programme
-      def self.edition_to_hash(edition,options={})
-        attrs = edition.programme.as_json(:only => [:audiences, :slug, :tags, :updated_at, :section, :related_items])
-        attrs.merge!(edition.as_json(:only => [:title, :alternative_title, :overview]))
-        attrs['parts'] = edition.parts.sort_by(&:order).collect { |p| p.as_json(:only => [:slug, :title, :body]).merge('number' => p.order) }
-        attrs['type'] = 'programme'
-        attrs
-      end
-    end
 
+    end
   end
 end
