@@ -8,13 +8,17 @@ class LocalTransaction < Publication
   validates_presence_of :lgsl, :on => :create
 
   set_callback :validation, :before do |local_transaction|
-    unless local_transaction.persisted?
+    unless local_transaction.persisted? or lgsl_code.blank?
       local_transaction.lgsl = LocalTransactionsSource.find_current_lgsl(local_transaction.lgsl_code)
     end
   end
 
   def self.edition_class
     LocalTransactionEdition
+  end
+
+  def search_format
+    "transaction"
   end
 
   def verify_snac(snac)
