@@ -1,12 +1,14 @@
 class Programme < Publication
   embeds_many :editions, :class_name => 'ProgrammeEdition', :inverse_of => :programme
 
+  include Parted
+
   DEFAULT_PARTS = [
-	{:title => "Overview", :slug => "overview"},
-	{:title => "What you'll get", :slug => "what-youll-get"},
-	{:title => "Eligibility", :slug => "eligibility"},
-	{:title => "How to claim", :slug => "how-to-claim"},
-	{:title => "Further information", :slug => "further-information"},
+  {:title => "Overview", :slug => "overview"},
+  {:title => "What you'll get", :slug => "what-youll-get"},
+  {:title => "Eligibility", :slug => "eligibility"},
+  {:title => "How to claim", :slug => "how-to-claim"},
+  {:title => "Further information", :slug => "further-information"},
   ]
 
   def self.edition_class
@@ -16,10 +18,9 @@ class Programme < Publication
   def create_first_edition
     unless self.persisted? or self.editions.any?
       self.editions << self.class.edition_class.new(:title => self.name)
-     DEFAULT_PARTS.each { |part|
+      DEFAULT_PARTS.each { |part|
         self.editions.first.parts.build(:title => part[:title],:slug => part[:slug], :body => " ")
       }
-      calculate_statuses
     end
   end
 end
