@@ -7,8 +7,9 @@ class TransactionEdition < Edition
   field :will_continue_on,  :type => String
   field :link,              :type => String
   field :more_information,  :type => String
+  field :alternate_methods,  :type => String
 
-  @fields_to_clone = [:introduction, :will_continue_on, :link, :more_information, :minutes_to_complete, :uses_government_gateway, :expectation_ids]
+  @fields_to_clone = [:introduction, :will_continue_on, :link, :more_information, :alternate_methods, :minutes_to_complete, :uses_government_gateway, :expectation_ids]
 
   def container
     self.transaction
