@@ -18,7 +18,7 @@ module Publisher
     # -- all .rb files in that directory are automatically loaded.
 
     # Custom directories with classes and modules you want to be autoloadable.
-    # config.autoload_paths += %W(#{config.root}/extras)
+    config.autoload_paths += %W(#{config.root}/lib #{config.root}/app/presenters)
 
     config.generators do |g|
         g.orm :mongoid
@@ -51,3 +51,8 @@ module Publisher
     config.filter_parameters += [:password]
   end
 end
+
+require 'open-uri'
+require 'builder'
+
+Object.send :include, Pethau::InitializeWith
