@@ -1,6 +1,8 @@
 Publisher::Application.configure do
   # Settings specified here will take precedence over those in config/environment.rb
 
+  config.middleware.use "Graylog2Exceptions", { :hostname => 'graylog.cluster', :port => '12201', :facility => 'publisher'}
+
   # The production environment is meant for finished, "live" apps.
   # Code is not reloaded between requests
   config.cache_classes = true
@@ -19,10 +21,11 @@ Publisher::Application.configure do
   # just comment this out and Rails will serve the files
 
   # See everything in the log (default is :info)
-  # config.log_level = :debug
+  config.log_level = :info
 
   # Use a different logger for distributed setups
   # config.logger = SyslogLogger.new
+  config.logger = GELF::Logger.new("graylog.cluster", "12201", max_size = 'WAN', { :facility => "publisher" })
 
   # Use a different cache store in production
   # config.cache_store = :mem_cache_store
@@ -30,9 +33,6 @@ Publisher::Application.configure do
   # Disable Rails's static asset server
   # In production, Apache or nginx will already do this
   config.serve_static_assets = false
-
-  # Enable serving of images, stylesheets, and javascripts from an asset server
-  # config.action_controller.asset_host = "http://assets.example.com"
 
   # Disable delivery errors, bad email addresses will be ignored
   # config.action_mailer.raise_delivery_errors = false
@@ -46,12 +46,12 @@ Publisher::Application.configure do
 
   # Send deprecation notices to registered listeners
   config.active_support.deprecation = :notify
-  
-  if Rails.env.production?
-    config.middleware.insert 0,  Slimmer::App, :template_host => "/data/vhost/static.alpha.gov.uk/current/public/templates"
-  else
-    config.middleware.insert 0,  Slimmer::App, :template_host => "/data/vhost/static.#{Rails.env}.alphagov.co.uk/current/public/templates"
-  end
-  
+
+  config.action_controller.asset_host = Proc.new { |source|
+    source =~ /publisher-assets/ ? nil : Plek.current.find('assets')
+  }
+  config.middleware.use Slimmer::App
+
   config.action_mailer.default_url_options = { :host => "www.gov.uk" }
+  # config.action_mailer.delivery_method = :ses
 end
