@@ -32,7 +32,7 @@ Publisher::Application.configure do
 
   # Print deprecation notices to the stderr
   config.active_support.deprecation = :stderr
-  
-  config.middleware.insert 0,  Slimmer::App, :template_host => "#{config.root}/public/templates"
+
+  config.middleware.use Slimmer::App, :asset_host => "http://static.preview.alphagov.co.uk"
   config.action_mailer.default_url_options = { :host => "example.com" }
 end
