@@ -1,6 +1,6 @@
 unless Rails.env.development? or Rails.env.test?
-  Publisher::Application.config.middleware.use ExceptionNotifier,
-    :email_prefix => "[Publisher] ",
+  Rails.application.config.middleware.use ExceptionNotifier,
+    :email_prefix => "[#{Rails.application.to_s.split('::').first}] ",
     :sender_address => %{"Winston Smith-Churchill" <winston@alphagov.co.uk>},
-    :exception_recipients => %w{dev@alphagov.co.uk}
+    :exception_recipients => %w{govuk-dev@digital.cabinet-office.gov.uk}
 end
