@@ -1,42 +1,32 @@
 Publisher::Application.routes.draw do
 
   namespace :admin do
-    resources :transactions do
-      post :progress, :on => :member
-      resources :editions
-    end
-    
-    resources :places do
-      post :progress, :on => :member
-      resources :editions
-    end
-    
-    resources :answers do
-      post :progress, :on => :member
-      resources :editions
-    end
-    
-    resources :guides do
-      post :progress, :on => :member
-      resources :editions
-    end
-    
-    resources :programmes do
-      post :progress, :on => :member
-      resources :editions
+    resources :notes
+    resources :expectations, :except => [:edit, :update, :destroy]
+    [ :transactions,
+      :places,
+      :answers,
+      :guides,
+      :programmes,
+      :local_transactions,
+    ].each do |r|
+      resources r do
+        resources :editions, :only => [:create, :update, :destroy] do
+          post :progress, :on => :member
+          post :start_work, :on => :member
+          member do
+            post 'skip_fact_check'
+          end
+        end
+      end
     end
 
-    resources :local_transactions do
-      post :progress, :on => :member
-      resources :editions
-    end
-    
-    match 'google_insight' => 'seo#google_insight'
-    
+    resources :publications
+    match 'overview' => 'overview#index'
     root :to => 'root#index'
   end
-  
-  resources :publications, :only => [:show,:index]
+
+  resources :publications, :only => :show
   resources :local_transactions, :only => :show do
     member do
       get '/all(.:format)', :to => :all
