@@ -1,19 +1,26 @@
 require 'test_helper'
 
-
 class Admin::ProgrammesControllerTest < ActionController::TestCase
 
   setup do
+    panopticon_has_metadata(
+      "id" => "12345",
+      "name" => "Test",
+      "slug" => "test"
+    )
     login_as_stub_user
-    without_panopticon_validation do
-      @programme = Programme.create(:name => "test", :slug=>"test")
-    end
+    @programme = Programme.create(:name => "test", :slug=>"test", :panopticon_id => 12345)
   end
 
   test "programmes index redirects to root" do
     get :index
     assert_response :redirect
     assert_redirected_to(:controller => "root", "action" => "index")
+  end
+
+  test "requesting a publication that doesn't exist returns a 404" do
+    get :show, :id => '4e663834e2ba80480a0000e6'
+    assert_response 404
   end
 
   test "we can view a programme" do
@@ -31,7 +38,8 @@ class Admin::ProgrammesControllerTest < ActionController::TestCase
   end
 
   test "can't destroy published programme" do
-    @programme.publish(@programme.editions.first, "test note")
+    @programme.editions.first.state = 'ready'
+    @programme.editions.first.publish
     assert !@programme.can_destroy?
     @programme.save!
     assert_difference('Programme.count', 0) do
