@@ -1,82 +1,46 @@
-require 'test_helper'
-require 'capybara/rails'
-
-class MockPanopticon
-  def initialize(app)
-    @app = app
-  end
-
-  def call(env)
-    if env['PATH_INFO'] == '/publications/slugs.json'
-      return [ 404, {}, "Slug Not Found" ]
-    else
-      @app.call(env)
-    end
-  end
-end
-
-Capybara.default_driver = :selenium
-Capybara.server_port = 4000
-Capybara.app = Rack::Builder.new do 
-  map "/" do
-    use MockPanopticon
-    run Capybara.app
-  end
-end
+require 'integration_test_helper'
 
 class AddingPartsToGuidesTest < ActionDispatch::IntegrationTest
-  include Capybara::DSL
-  
-  def teardown
-    DatabaseCleaner.clean
-  end
-  
+
   test "Publishing a guide" do
-     without_panopticon_validation do 
-     
-       # This isn't right, really need a way to run actions when
-     # logged in as particular users without having Signonotron running.
-     #
-     User.create(:uid=>"ADADS",:name=>"T Est",:email=>"test@example.com")
-   
-     random_name = (0...8).map{65.+(rand(25)).chr}.join + " GUIDE"
+    setup_users
 
-     visit    "/admin"
-     click_on 'Add new guide'
-     fill_in  'Name', :with => random_name
-     fill_in  'Slug',  :with => 'test-guide'
-     click_on 'Create Guide'
-     
-     click_on 'Untitled part'
-     within :css, '#parts div.part:first-of-type' do
-       fill_in 'Title' , :with => 'Part One'
-       fill_in 'Body', :with => 'Body text'
-       fill_in 'Slug', :with => 'part-one'
-     end
-     click_on 'Add new part'
-     within :css, '#parts div.part:nth-of-type(2)' do
-       fill_in 'Title' , :with => 'Part Two'
-       fill_in 'Body', :with => 'Body text'
-       fill_in 'Slug', :with => 'part-two'
-     end
-     within(:css, '#guide-controls') { click_on 'Save' }
+    random_name = (0...8).map{65.+(rand(25)).chr}.join + " GUIDE"
 
-     click_on 'Add new part'
-     within :css, '#parts div.part:nth-of-type(3)' do
-       fill_in 'Title' , :with => 'Part Three'
-       fill_in 'Body', :with => 'Body text'
-       fill_in 'Slug', :with => 'part-three'
-     end
-     within(:css, '#guide-controls') { click_on 'Save' }
-   
-     assert_equal 3, all(:css, '#parts > div.part').length
+    guide = Guide.new :name => random_name, :slug => 'test-guide', :panopticon_id => 2356
+    guide.save!
+    guide.editions.first.update_attribute(:state, 'draft')
 
-     visit "/admin"
-     
-     within(:css, '#new') {
-        assert page.has_content? random_name 
-     }
+    panopticon_has_metadata("id" => '2356')
 
-     end
-   end
+    visit    "/admin/guides/#{guide.to_param}"
+
+    click_on 'Untitled part'
+    within :css, '#parts div.part:first-of-type' do
+      fill_in 'Title', :with => 'Part One'
+      fill_in 'Body',  :with => 'Body text'
+      fill_in 'Slug',  :with => 'part-one'
+    end
+
+    click_on 'Add new part'
+    within :css, '#parts div.part:nth-of-type(2)' do
+      fill_in 'Title', :with => 'Part Two'
+      fill_in 'Body',  :with => 'Body text'
+      fill_in 'Slug',  :with => 'part-two'
+    end
+    within(:css, '#publication-controls') { click_on 'Save' }
+
+    click_on 'Add new part'
+    within :css, '#parts div.part:nth-of-type(3)' do
+      fill_in 'Title', :with => 'Part Three'
+      fill_in 'Body',  :with => 'Body text'
+      fill_in 'Slug',  :with => 'part-three'
+    end
+    within(:css, '#publication-controls') { click_on 'Save' }
+
+    assert_equal 3, all(:css, '#parts > div.part').length
+
+    visit "/admin?filter=all&list=drafts"
+    assert page.has_content? random_name
+  end
 end
