@@ -1,10 +1,6 @@
 require 'test_helper'
 
 class LocalTransactionTest < ActiveSupport::TestCase
-  def setup
-    PanopticonSlugValidator.any_instance.stubs(:validate).returns(true)
-  end
-
   def lgsl
     @lgsl ||= LocalTransactionsSource::Lgsl.create(code: "1")
   end
@@ -15,22 +11,17 @@ class LocalTransactionTest < ActiveSupport::TestCase
   alias_method :authority, :create_authority
 
   test "looks up the LGSL before validating a new record" do
-    without_panopticon_validation do
-      LocalTransactionsSource.expects(:find_current_lgsl).with("1").returns(lgsl)
-
-      lt = LocalTransaction.new(lgsl_code: "1", name: "Transaction", slug: "slug")
-      assert lt.valid?
-    end
+    LocalTransactionsSource.expects(:find_current_lgsl).with("1").returns(lgsl)
+    lt = LocalTransaction.new(lgsl_code: "1", name: "Transaction", slug: "slug")
+    assert lt.valid?
   end
 
   test "doesn't bother looking up the LGSL before validating an existing record" do
-    without_panopticon_validation do
-      LocalTransactionsSource.expects(:find_current_lgsl).never
+    LocalTransactionsSource.expects(:find_current_lgsl).never
 
-      lt = LocalTransaction.new(lgsl_code: "1", name: "Transaction", slug: "slug")
-      lt.save(validate: false)
-      assert lt.valid?
-    end
+    lt = LocalTransaction.new(lgsl_code: "1", name: "Transaction", slug: "slug")
+    lt.save(validate: false)
+    assert lt.valid?
   end
 
   test "can verify whether an authority provides the transaction service, given its SNAC" do
@@ -47,5 +38,10 @@ class LocalTransactionTest < ActiveSupport::TestCase
     create_authority("45UB")
 
     assert !lt.verify_snac("00BC")
+  end
+
+  test "local transaction reports itself as a transaction for search" do
+    lt = LocalTransaction.new(lgsl_code: "1", name: "Transaction", slug: "slug")
+    assert_equal "transaction", lt.search_format
   end
 end
